requirements.txt
discord.py
asyncpg
python-dotenv
sqlalchemy
greenlet
